Refactored to split off ModalHeader into own file/component - still too large and may need to split again. Header is also missing updated styling for the reviews component (change the size and padding to align with rest of header)
