+++ client/src/components/Modal.jsx
@@ -1,0 +1,56 @@
+import React from 'react';
+import styled from 'styled-components';
+import ModalHeader from './ModalHeader.jsx';
+
+const Modal = styled.div`
+  position: fixed;
+  top: 10%;
+  width: 50vw;
+  height: 70%;
+  background-color: white;
+  box-shadow: 0 0 3px black;
+  z-index: 1;
+  align-self: center;
+`
+
+const ContentContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  padding: 48px;
+  box-sizing: border-box;
+`
+
+const Exit = styled.div`
+  position: absolute;
+  z-index: 1;
+  top: 0;
+  right: 0;
+  text-align: center;
+  cursor: pointer;
+  padding: 15px;
+`
+
+const Category = styled.span`
+  font-size: 12px;
+  font-weight: 500;
+  color: gray;
+`
+
+const Name = styled.span`
+  font-size: 24px;
+  font-weight: 500;
+`
+
+const TourModal = (props) => (
+  <Modal>
+    <Exit> <i className="fas fa-times" onClick={props.toggleModal}></i> </Exit>
+    <ContentContainer>
+      <Category>{props.item.categories_name}</Category>
+      <Name>{props.item.name}</Name>
+      <ModalHeader item={props.item}/>
+    </ContentContainer>
+  </Modal>
+)
+
+export default TourModal;
